client-api: Move AuthData to its own module
Given that it includes a lot of types, and more are coming.
